@@ -1,9 +1,11 @@
 import * as React from "react"
-import * as ReactDOM from "react-dom"
+import * as ReactDOM from "react-dom/client"
 import "./index.css"
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faArrowRight } from "@fortawesome/free-solid-svg-icons"
+
+document.title = "Home"
 
 const App = () => {
     return <>
@@ -24,5 +26,5 @@ const App = () => {
     </>
 }
 
-const root = document.getElementById("root")
-ReactDOM.render(<App />, root)
+const root = ReactDOM.createRoot(document.getElementById("root"))
+root.render(<App />)
